Add file & checksum utils
